[r3163] Устраняем некоторые предупреждения компилятора: 1. Используем явное приведение целых типов с большей разрядности к меньшей. 2. В функции swscanf_s задаем размер буфера равный 1 при чтении одиночных символов "%c". 3. MpegSplitter: отключаем макрос MARKER перед его новым определением. 4. MPCVideoDec: отключаем предупреждение С4005 для макроса MAX_SLICES при включении файлов ffmpeg/libavcodec/. 5. DSUtil: устранена ошибка в функции ReadDisplay(), когда указатель приводился к типу UINT16.

+++ src/DSUtil/DSUtil.cpp
@@ -2925,7 +2925,7 @@ HRESULT CreateAVCfromH264(CMediaType* mt)
 	while (Nalu.ReadNext()) {
 		if (Nalu.GetType() == NALU_TYPE_SPS || Nalu.GetType() == NALU_TYPE_PPS) {
 			size_t nalLength = Nalu.GetDataLength();
-			*(dst + dstSize++) = nalLength >> 8;
+			*(dst + dstSize++) = (BYTE)(nalLength >> 8);
 			*(dst + dstSize++) = nalLength & 0xff;
 
 			memcpy(dst + dstSize, Nalu.GetDataBuffer(), Nalu.GetDataLength());
